backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
mysqlclient>=2.2,<3.0
redis>=5.0,<6.0
celery[redis]>=5.4,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
django_fsm>=3.0,<4.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
requests>=2.31,<3.0
openpyxl>=3.1,<4.0
drf-yasg>=1.21,<2.0
gunicorn>=22.0,<23.0
oss2>=2.18,<3.0
